Add const constructor
