+++ super_editor/lib/src/super_reader/reader_context.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
 import 'package:super_editor/src/core/document.dart';
-import 'package:super_editor/src/core/document_composer.dart';
 import 'package:super_editor/src/core/document_layout.dart';
 import 'package:super_editor/src/core/document_selection.dart';
 import 'package:super_editor/src/default_editor/document_scrollable.dart';
